Plugins: Make matplotlib have its own plugin
* That one can be always enabled, as it's being harmless if
  matplotlib is not used and even if it's not really used.

* Numpy plugin now warns about asking to not include matplotlib,
  that is no longer its business.

* Solved TODO about using proper query function for values, this
  had been code factored out of the matplotlib scanning code, but
  should be used there of course too, as it removes a lot of noise
  from it.

* Also namespace info names per plugin, seems safer for copy paste
  code.

* Use directory copy for matplotlib which makes it a lot less
  verbose.
